Use a hardcoded Persona version that's not tied to OUR version.
Apparently, instead of forking Persona theme, we have a post-install
hook that DOWNLOADS the Persona theme and copies all of its files except
the ones we want to override. We also import Persona theme in the
theme.less file. Unfortunately, the version of Persona installed was
chosen to be the same as THIS package's version, which creates...
trouble, if any changes are made to this package. Use a hardcoded
version that represents the version of Persona we're basing the theme
off of.